Add comment about gevent option not supported
Following latest test we discovered that the agent was not collecting data when -k gevent option was used for gevent workers. This is a limitation for now, we have not investigated further into the issue and have no solution to propose yet.
Adding a comment in this readme to notify customers about it.

Also fixed a typo for generate-traffic.sh script name.